apps/capture/requirements.txt
pygetwindow
pyautogui
opencv-python
ffmpeg-python
numpy
mss
python-socketio[client]